requirements-dev.txt
asgiref==3.2.10
certifi==2020.6.20
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-cors-headers==3.4.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
drf-writable-nested==0.6.3
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
